requirements.txt
tensorflow-cpu
python-whois
fastapi
uvicorn
bs4
